requirements.txt
#General module
pipenv

#Dev
hikari==2.0.0.dev109
hikari-lightbulb==2.2.2
python-dotenv==0.20.0
requests==2.28.1